Switched the name of the beta filehandler method and the original version.
