+++ Contacts/contacts.txt
@@ -1,3 +1,2 @@
 bob, 123
-bob, 345
-joyce, 123
+joe, 234
